+++ dmr/openapi/views/yaml.py
@@ -3,7 +3,15 @@ from typing import ClassVar
 from django.http import HttpRequest, HttpResponse
 
 from dmr.openapi.views.base import OpenAPIView
-from dmr.plugins.msgspec.yaml import yaml_dumps
+
+try:
+    import yaml
+except ImportError:
+    print(  # noqa: WPS421
+        'Looks like `pyyaml` is not installed, '
+        "consider using `pip install 'django-modern-rest[openapi]'`",
+    )
+    raise
 
 
 class OpenAPIYamlView(OpenAPIView):
@@ -11,7 +19,7 @@ class OpenAPIYamlView(OpenAPIView):
     View for returning the OpenAPI schema as YAML.
 
     This view mirrors :class:`~dmr.openapi.views.json.OpenAPIJsonView`,
-    but renders the converted schema using ``msgspec.yaml``.
+    but renders the converted schema using ``pyyaml``.
     Produces a YAML representation of the :class:`~dmr.openapi.objects.OpenAPI`
     specification that can be used by API documentation tools
     and client code generators.
@@ -22,6 +30,6 @@ class OpenAPIYamlView(OpenAPIView):
     def get(self, request: HttpRequest) -> HttpResponse:
         """Render the OpenAPI schema as YAML response."""
         return HttpResponse(
-            content=yaml_dumps(self.schema.convert()),
+            content=yaml.safe_dump(self.schema.convert()),
             content_type=self.content_type,
         )