+++ CHANGELOG.md
@@ -29,6 +29,7 @@ No breaking changes will be made.
 - Added an option to override `exclude_semantic_responses`
   and `no_validate_http_spec` settings with `None`
 - Added a new way to resolve annotations for controllers, #787
+- Added `yaml` view for OpenAPI schema, #745
 
 ### Fixes